remove CardTableController and distribute logic to proper layers
Move card table endpoints to CardController, extract specification
methods to CardSpecifications, and move business logic to CardService.
