+++ server/pom.xml
@@ -165,6 +165,10 @@
               <groupId>org.projectlombok</groupId>
               <artifactId>lombok</artifactId>
             </path>
+            <path>
+              <groupId>org.hibernate.orm</groupId>
+              <artifactId>hibernate-jpamodelgen</artifactId>
+            </path>
           </annotationProcessorPaths>
         </configuration>
       </plugin>
