+++ server/src/main/java/io/github/mucsi96/learnlanguage/repository/CardRepository.java
@@ -3,9 +3,11 @@ package io.github.mucsi96.learnlanguage.repository;
 import io.github.mucsi96.learnlanguage.entity.Card;
 import io.github.mucsi96.learnlanguage.entity.Source;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -19,4 +21,10 @@ public interface CardRepository
 
     @Modifying
     void deleteBySource(Source source);
+
+    @Query(value = "SELECT * FROM learn_language.cards ORDER BY random()", nativeQuery = true)
+    List<Card> findRandomCards(Pageable pageable);
+
+    @Query("SELECT c.source.id, COUNT(c) FROM Card c GROUP BY c.source")
+    List<Object[]> countCardsBySourceGroupBySource();
 }
