+++ server/src/main/java/io/github/mucsi96/learnlanguage/repository/LearningPartnerRepositoryCustomImpl.java
@@ -1,6 +1,7 @@
 package io.github.mucsi96.learnlanguage.repository;
 
 import io.github.mucsi96.learnlanguage.entity.LearningPartner;
+import io.github.mucsi96.learnlanguage.entity.LearningPartner_;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaUpdate;
@@ -17,8 +18,8 @@ public class LearningPartnerRepositoryCustomImpl implements LearningPartnerRepos
         final CriteriaUpdate<LearningPartner> update = cb.createCriteriaUpdate(LearningPartner.class);
         final Root<LearningPartner> root = update.from(LearningPartner.class);
 
-        update.set(root.<Boolean>get("isActive"), false);
-        update.where(cb.notEqual(root.get("id"), id));
+        update.set(root.get(LearningPartner_.isActive), false);
+        update.where(cb.notEqual(root.get(LearningPartner_.id), id));
 
         entityManager.createQuery(update).executeUpdate();
     }
