+++ server/src/main/java/io/github/mucsi96/learnlanguage/repository/ModelUsageLogRepository.java
@@ -3,6 +3,7 @@ package io.github.mucsi96.learnlanguage.repository;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import io.github.mucsi96.learnlanguage.entity.ModelUsageLog;
@@ -10,10 +11,16 @@ import io.github.mucsi96.learnlanguage.model.ModelType;
 import io.github.mucsi96.learnlanguage.model.OperationType;
 
 @Repository
-public interface ModelUsageLogRepository
-        extends JpaRepository<ModelUsageLog, Long>, ModelUsageLogRepositoryCustom {
+public interface ModelUsageLogRepository extends JpaRepository<ModelUsageLog, Long> {
     List<ModelUsageLog> findAllByOrderByCreatedAtDesc();
     List<ModelUsageLog> findByModelTypeOrderByCreatedAtDesc(ModelType modelType);
     List<ModelUsageLog> findByOperationTypeOrderByCreatedAtDesc(OperationType operationType);
     List<ModelUsageLog> findByResponseContent(String responseContent);
+
+    @Query("SELECT m.modelName, COUNT(m), COUNT(m.rating), "
+            + "COALESCE(AVG(m.rating), 0.0), "
+            + "COALESCE(SUM(m.costUsd), 0) "
+            + "FROM ModelUsageLog m GROUP BY m.modelName "
+            + "ORDER BY COALESCE(AVG(m.rating), 0.0) DESC")
+    List<Object[]> getModelSummary();
 }
