--- server/src/main/java/io/github/mucsi96/learnlanguage/repository/ModelUsageLogRepositoryCustom.java
@@ -1,7 +1,0 @@
-package io.github.mucsi96.learnlanguage.repository;
-
-import java.util.List;
-
-public interface ModelUsageLogRepositoryCustom {
-    List<Object[]> getModelSummary();
-}
