--- server/src/main/java/io/github/mucsi96/learnlanguage/repository/ReviewLogRepositoryCustom.java
@@ -1,9 +1,0 @@
-package io.github.mucsi96.learnlanguage.repository;
-
-import io.github.mucsi96.learnlanguage.entity.ReviewLog;
-
-import java.util.List;
-
-public interface ReviewLogRepositoryCustom {
-    List<ReviewLog> findLatestReviewsByCardIds(List<String> cardIds);
-}
