--- server/src/main/java/io/github/mucsi96/learnlanguage/repository/StudySessionRepositoryCustom.java
@@ -1,7 +1,0 @@
-package io.github.mucsi96.learnlanguage.repository;
-
-import java.time.LocalDateTime;
-
-public interface StudySessionRepositoryCustom {
-    void deleteByCreatedAtBefore(LocalDateTime cutoff);
-}
